pass rocksdb ut when isMac()

+++ store/src/test/java/org/apache/rocketmq/store/queue/CombineConsumeQueueStoreTest.java
@@ -24,6 +24,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.TimeUnit;
 import org.apache.commons.io.FileUtils;
+import org.apache.rocketmq.common.MixAll;
 import org.apache.rocketmq.common.TopicConfig;
 import org.apache.rocketmq.common.UtilAll;
 import org.apache.rocketmq.common.attribute.CQType;
@@ -59,12 +60,18 @@ public class CombineConsumeQueueStoreTest extends QueueTestBase {
 
     @Before
     public void init() throws Exception {
+        if (MixAll.isMac()) {
+            return;
+        }
         this.topicConfigTableMap = new ConcurrentHashMap<>();
         messageStoreConfig = new MessageStoreConfig();
     }
 
     @After
     public void destroy() {
+        if (MixAll.isMac()) {
+            return;
+        }
         messageStore.shutdown();
         messageStore.destroy();
 
@@ -74,6 +81,9 @@ public class CombineConsumeQueueStoreTest extends QueueTestBase {
 
     @Test(expected = IllegalArgumentException.class)
     public void CombineConsumeQueueStore_EmptyLoadingCQTypes_ThrowsException() throws Exception {
+        if (MixAll.isMac()) {
+            return;
+        }
         messageStore = (DefaultMessageStore) createMessageStore(null, false, topicConfigTableMap, messageStoreConfig);
 
         messageStoreConfig.setCombineCQLoadingCQTypes("");
@@ -82,6 +92,9 @@ public class CombineConsumeQueueStoreTest extends QueueTestBase {
 
     @Test
     public void CombineConsumeQueueStore_InitializesConsumeQueueStore() throws Exception {
+        if (MixAll.isMac()) {
+            return;
+        }
         messageStore = (DefaultMessageStore) createMessageStore(null, false, topicConfigTableMap, messageStoreConfig);
         {
             messageStoreConfig.setCombineCQLoadingCQTypes("default");
@@ -121,6 +134,9 @@ public class CombineConsumeQueueStoreTest extends QueueTestBase {
 
     @Test
     public void testIterator() throws Exception {
+        if (MixAll.isMac()) {
+            return;
+        }
         messageStoreConfig.setRocksdbCQDoubleWriteEnable(true);
         messageStore = (DefaultMessageStore) createMessageStore(null, false, topicConfigTableMap, messageStoreConfig);
         messageStore.load();
@@ -201,6 +217,9 @@ public class CombineConsumeQueueStoreTest extends QueueTestBase {
 
     @Test
     public void testInitializeWithOffset() throws Exception {
+        if (MixAll.isMac()) {
+            return;
+        }
         final String path = createBaseDir();
         FileUtils.deleteDirectory(new File(path));
         topicConfigTableMap.put(topic, new TopicConfig(topic, 1, 1, PermName.PERM_WRITE | PermName.PERM_READ));
@@ -293,6 +312,9 @@ public class CombineConsumeQueueStoreTest extends QueueTestBase {
 
     @Test
     public void testVerifyAndInitOffsetForAllStore() throws Exception {
+        if (MixAll.isMac()) {
+            return;
+        }
         final String path = createBaseDir();
         topicConfigTableMap.put(topic, new TopicConfig(topic, 1, 1, PermName.PERM_WRITE | PermName.PERM_READ));
